delete v-header and move code to AppHeader
