+++ src/components/AppHeader.vue
@@ -10,6 +10,8 @@ import CloseIcon from '@/components/icons/CloseIcon.vue'
 const router = useRouter()
 const canShowBanner = ref(true)
 const hideHeader = ref(false)
+const showHelpMenu = ref(false)
+const showMenu = ref(false)
 
 let lastScrollTop = 0
 const scrollDelta = 10
@@ -96,14 +98,33 @@ function handleScroll(ev: any) {
         <CloseIcon color="#FFFFFF" />
       </span>
     </div>
-    <v-header
-      :logo-src="ArcanaLogo"
-      logo-alt="Arcana Logo"
-      :help-items="helpItems"
-      :mobile-menu-icon="MenuIcon"
-      :close-icon="BlueCloseIcon"
-      @logo-click="onLogoClick"
-    />
+    <header class="container flex" style="margin-bottom: 0">
+      <div class="logo" @click.stop="onLogoClick">
+        <img :src="ArcanaLogo" alt="Arcana Logo" />
+      </div>
+      <div class="position-relative flex flex-center">
+        <button class="help-button" @click.stop="showHelpMenu = !showHelpMenu">
+          Help
+        </button>
+        <button
+          class="mobile-menu-icon-button cursor-pointer"
+          @click.stop="showMenu = !showMenu"
+        >
+          <img v-if="showMenu" :src="BlueCloseIcon" alt="close icon" />
+          <img v-else :src="MenuIcon" alt="Menu icon" />
+        </button>
+        <ul v-if="showHelpMenu" class="help-menu-items position-absolute">
+          <li
+            v-for="helpItem in helpItems"
+            :key="helpItem.label"
+            class="cursor-pointer help-menu-item"
+            @click.stop="showHelpMenu = false"
+          >
+            {{ helpItem.label }}
+          </li>
+        </ul>
+      </div>
+    </header>
   </section>
 </template>
 
@@ -114,6 +135,102 @@ section {
   z-index: 1000;
   width: 100%;
   transition: transform 0.4s;
+}
+
+header {
+  position: relative;
+  display: flex;
+  justify-content: space-between;
+  padding: 1.5rem 0;
+  visibility: visible;
+  background: #1f1f1f;
+  transition: transform 0.6s;
+}
+
+.mobile-menu {
+  position: absolute;
+  top: 100%;
+  z-index: 10000;
+  width: 100%;
+  height: 100vh;
+  visibility: hidden;
+  background-color: #262626;
+  opacity: 0;
+  transition: opacity 0.6s, visibility 0.6s;
+}
+
+.mobile-menu-icon-button {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 24px;
+  height: 24px;
+  background-color: transparent;
+  border: none;
+  outline: none;
+}
+
+.hide-header {
+  transform: translateY(-100%);
+}
+
+.logo {
+  cursor: pointer;
+  transition: opacity 0.3s;
+}
+
+.logo:hover {
+  opacity: 0.6;
+}
+
+.help-button {
+  font-family: var(--font-body);
+  color: var(--primary);
+  cursor: pointer;
+  background: transparent;
+  border: none;
+  outline: none;
+}
+
+.help-menu-items {
+  top: 30px;
+  right: 20px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  width: 245px;
+  padding: 0;
+  padding: 1rem;
+  background-color: #1f1f1f;
+  border-radius: 10px;
+  box-shadow: -4px -5px 4px rgb(0 0 0 / 20%), 4px 5px 4px rgb(0 0 0 / 20%);
+}
+
+.help-menu-item {
+  width: 100%;
+  padding: 0.5rem;
+  font-family: var(--font-body);
+  color: var(--text-white);
+  list-style: none;
+  background-color: #1f1f1f;
+}
+
+@media only screen and (max-width: 767px) {
+  .help-button {
+    display: none;
+  }
+}
+
+@media only screen and (min-width: 768px) {
+  .mobile-menu-icon-button {
+    display: none;
+  }
+}
+
+@media only screen and (max-width: 1023px) {
+  header {
+    padding: 1rem 0;
+  }
 }
 
 .banner {
@@ -147,9 +264,5 @@ section {
   position: absolute;
   right: 2em;
   cursor: pointer;
-}
-
-.hide-header {
-  transform: translateY(-100%);
 }
 </style>
